wrote failing integration test

+++ ClearMeasure.Interview/ClearMeasure.AFA/FizzBuzzer.cs
@@ -19,7 +19,7 @@ namespace ClearMeasure.AFA
             this.divisorStrategies = divisorStrategies;
         }
 
-        public void Execute(int upperBound = 100)
+        public void Execute(long upperBound = 100)
         {
             for (int i = 1; i < upperBound + 1; i++)
             {
